Fixed SaveTextFile error
When PandQ.saveTextFile is executed, it saves to the file path that
saved as PandQ.file. When PandQ.file = null, an exception is thrown.
Now, in saveTextFile is executed, it first checks to see if there is a
file at all. If not, the user chooses a place to save.

in ViewAllQuestionsForm, it checks to see if any questions were even
selected, removing the error.

+++ StudyWizard/ViewAllQuestionsForm.cs
@@ -32,9 +32,12 @@ namespace StudyWizard
 
         private void btn_delete_Click(object sender, EventArgs e)
         {
-            pandQ.deleteQuestion(chdlsbx_questions.CheckedIndices);
-            chdlsbx_questions.ClearSelected();
-            chdlsbx_questions.DataSource = pandQ.viewAllQuestions();
+            if (chdlsbx_questions.CheckedIndices == null)
+            {
+                pandQ.deleteQuestion(chdlsbx_questions.CheckedIndices);
+                chdlsbx_questions.ClearSelected();
+                chdlsbx_questions.DataSource = pandQ.viewAllQuestions();
+            }
         }
 
         private void btn_edit_Click(object sender, EventArgs e)
